修复一个bug

+++ EroMangaManager.UWP/Views/FunctionChildPages/FindSameManga.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 using EroMangaManager.Core.Models;
 using EroMangaManager.UWP.Helpers;
@@ -49,17 +50,26 @@ namespace EroMangaManager.UWP.Views.FunctionChildPages
             var button = sender as Button;
             var manga = button.DataContext as MangaBook;
 
-            await Helpers.StorageHelper.DeleteSourceFile(manga);
+            var deleteResult= await Helpers.StorageHelper.DeleteSourceFile(manga);
 
-            mangaBookViewModel.DeleteStorageFileInRootObservable(manga);
-        }
+            if (deleteResult)
+            {
+                mangaBookViewModel.DeleteStorageFileInRootObservable(manga);
+            }
+            }
 
-        private async void OpenMangaClick(object sender, RoutedEventArgs e)
+            private async void OpenMangaClick(object sender, RoutedEventArgs e)
         {
             var button = sender as Button;
             var manga = button.DataContext as MangaBook;
 
             await WindowHelper.ShowNewReadPageWindow(manga);
+        }
+
+        private async void LauncherFolder(object sender, RoutedEventArgs e)
+        {
+            var mangaBook = (sender as MenuFlyoutItem).DataContext as MangaBook;
+            await Windows.System.Launcher.LaunchFolderPathAsync(mangaBook.FolderPath);
         }
     }
 }
